Split file size docs-truth guards
